refactor out step items
